Adicionei Badge para número de matches na Tela Inicial.
